refactor(test): split experimental memory tests into dedicated file
Move 6 [experimental]-tagged tests from memory-regressions.test.ts to
memory-regressions.experimental.test.ts. Add the new file to
EXPERIMENTAL_FILES in test.sh for file-level exclusion.

This completes the PR6 plan requirement of separating experimental
tests into a dedicated file rather than relying solely on name-pattern
regex filtering.
